[Fix] SideNav location
